# backend/docs/openapi.yaml
openapi: "3.0.3"
info:
  title: Bike Auction Platform API
  description: |
    RESTful API for the Bike Auction Platform.

    ## Authentication
    - **Register/Login** → Receive a short-lived JWT Access Token (15 min) in the response body
      and a long-lived Refresh Token (7 days) as an `httpOnly`, `SameSite=Strict`, `Secure` cookie.
    - **Authenticated requests** → Pass the Access Token in the `Authorization: Bearer <token>` header.
    - **Token refresh** → `POST /api/auth/refresh` reads the cookie and returns a new Access Token.
    - **Real-time bidding** → Uses WebSocket (Socket.io) with token passed in the connection handshake.
      See the architecture document for WebSocket event contracts.

    ## Monetary Values
    All monetary fields are represented in **cents** (integer) to avoid floating-point precision issues.
    Example: `$1,250.00` → `125000`.
  version: "1.0.0"
  contact:
    name: Bike Auction Engineering
    email: eng@bikeauction.dev
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.bikeauction.dev
    description: Production

# ============================================================================
# Security Schemes
# ============================================================================
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived access token (15 minutes). Obtain via POST /api/auth/login.

  # ==========================================================================
  # Reusable Schemas
  # ==========================================================================
  schemas:
...
    MotorcycleCreate:
      type: object
      required: [title, make, model, year, mileageKm, condition]
      properties:
        title:
          type: string
          maxLength: 300
          example: "2023 Ducati Panigale V4 S — Pristine Condition"
        make:
          type: string
          maxLength: 100
          example: Ducati
        model:
          type: string
          maxLength: 100
          example: Panigale V4 S
        year:
          type: integer
          minimum: 1900
          maximum: 2100
          example: 2023
        vin:
          type: string
          maxLength: 17
          example: ZDM14BKW1PB000123
        mileageKm:
          type: integer
          minimum: 0
          example: 3200
        engineCc:
          type: integer
          example: 1103
        condition:
          type: string
          enum: [excellent, good, fair, poor, salvage]
          example: excellent
        description:
          type: string
          example: "One owner, full service history, never dropped."
        photoUrls:
          type: array
          items:
            type: string
            format: uri
          example:
            - https://cdn.bikeauction.dev/photos/ducati-front.jpg
            - https://cdn.bikeauction.dev/photos/ducati-side.jpg
        thumbnailUrl:
          type: string
          format: uri
...
    Motorcycle:
      allOf:
        - $ref: "#/components/schemas/MotorcycleCreate"
        - type: object
          properties:
            id:
              type: string
              format: uuid
            sellerId:
              type: string
              format: uuid
            listingStatus:
              type: string
              enum: [draft, pending_approval, approved, rejected]
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time

    # --- Auction ---
...
    Auction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        motorcycleId:
          type: string
          format: uuid
        sellerId:
          type: string
          format: uuid
        startingPriceCents:
          type: integer
        reservePriceCents:
          type: integer
          nullable: true
        bidIncrementCents:
          type: integer
        currentHighestBidCents:
          type: integer
          description: "Denormalized highest bid for fast reads."
        currentHighestBidderId:
          type: string
          format: uuid
          nullable: true
        totalBids:
          type: integer
        startsAt:
          type: string
          format: date-time
        originalEndTime:
          type: string
          format: date-time
        currentEndTime:
          type: string
          format: date-time
          description: "May differ from originalEndTime due to soft-close extensions."
        softCloseWindowSeconds:
          type: integer
        softCloseExtensionSeconds:
          type: integer
        status:
          type: string
          enum: [scheduled, active, soft_close, ended, reserve_not_met, sold, cancelled]
        version:
          type: integer
          description: "Optimistic concurrency version. Incremented on every bid."
        motorcycle:
          $ref: "#/components/schemas/Motorcycle"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # --- Bid ---
    Bid:
      type: object
      properties:
        id:
          type: string
          format: uuid
        auctionId:
          type: string
          format: uuid
        bidderId:
          type: string
          format: uuid
        bidderUsername:
          type: string
          description: "Populated via join for display."
        amountCents:
          type: integer
        status:
          type: string
          enum: [valid, outbid, rejected, winning]
        correlationId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time

    # --- Pagination ---
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        totalItems:
          type: integer
          example: 142
        totalPages:
          type: integer
          example: 8

  # ==========================================================================
  # Reusable Parameters
  # ==========================================================================
  parameters:
    PathId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    QueryPage:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    QueryLimit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20

# ============================================================================
# Paths
# ============================================================================
paths:
...
  /api/users/me/bids:
    get:
      tags: [Users]
      summary: Get my bid history across all auctions
      operationId: getMyBids
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/QueryPage"
        - $ref: "#/components/parameters/QueryLimit"
        - name: status
          in: query
          schema:
            type: string
            enum: [valid, outbid, rejected, winning]
      responses:
        "200":
          description: Paginated list of the current user's bids
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Bid"
                  meta:
                    $ref: "#/components/schemas/PaginationMeta"
  /api/users/me/auctions:
    get:
      tags: [Users]
      summary: Get my auctions (seller dashboard)
      operationId: getMyAuctions
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/QueryPage"
        - $ref: "#/components/parameters/QueryLimit"
        - name: status
          in: query
          schema:
            type: string
            enum: [scheduled, active, soft_close, ended, reserve_not_met, sold, cancelled]
      responses:
        "200":
          description: Paginated list of the seller's auctions
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Auction"
                  meta:
                    $ref: "#/components/schemas/PaginationMeta"
  /api/users/me/motorcycles:
    get:
      tags: [Users]
      summary: Get my motorcycle listings (seller dashboard)
      operationId: getMyMotorcycles
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/QueryPage"
        - $ref: "#/components/parameters/QueryLimit"
        - name: listingStatus
          in: query
          schema:
            type: string
            enum: [draft, pending_approval, approved, rejected]
      responses:
        "200":
          description: Paginated list of the seller's motorcycle listings
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Motorcycle"
                  meta:
                    $ref: "#/components/schemas/PaginationMeta"
  /api/users/me/notifications:
    get:
      tags: [Users]
      summary: Get my notifications
      operationId: getMyNotifications
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/QueryPage"
        - $ref: "#/components/parameters/QueryLimit"
        - name: unreadOnly
          in: query
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Paginated notifications list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                        type:
                          type: string
                          enum: [outbid, auction_won, auction_ending, auction_cancelled]
                        auctionId:
                          type: string
                          format: uuid
                        payload:
                          type: object
                        isRead:
                          type: boolean
                        deliveredVia:
                          type: string
                        createdAt:
                          type: string
                          format: date-time
                  meta:
                    $ref: "#/components/schemas/PaginationMeta"
  /api/users/me/notifications/{id}/read:
    post:
      tags: [Users]
      summary: Mark a notification as read
      operationId: markNotificationRead
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Notification marked as read
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Notification marked as read

  # --------------------------------------------------------------------------
  # Health
  # --------------------------------------------------------------------------
...